cookbook/pocketflow-parallel-batch-flow/requirements.txt
pocketflow
Pillow>=10.0.0  # 用于图像处理
numpy>=1.24.0   # 用于图像数组操作